+++ packages/desktop/views/components/index.ts
@@ -1,5 +1,6 @@
 export { default as Background } from './Background.svelte'
 export { default as BgGradient } from './BgGradient.svelte'
+export { default as ButtonTile } from './ButtonTile.svelte'
 export { default as LoggedOutLayout } from './LoggedOutLayout.svelte'
 export { default as OnboardingLayout } from './OnboardingLayout.svelte'
 export { default as Particles } from './Particles.svelte'
