+++ packages/desktop/views/dashboard/developer/components/MintNativeTokenButton.svelte
@@ -1,9 +1,10 @@
 <script lang="ts">
+    import { IconName } from '@bloomwalletio/ui'
     import { selectedAccount } from '@core/account/stores'
     import { localize } from '@core/i18n'
     import { resetMintTokenDetails } from '@core/wallet'
     import { closePopup, openPopup, PopupId } from '@desktop/auxiliary/popup'
-    import { OnboardingButton } from '@ui'
+    import { ButtonTile } from '../../../components'
 
     $: hasAliases = $selectedAccount.balances?.aliases.length > 0
 
@@ -32,8 +33,9 @@
     }
 </script>
 
-<OnboardingButton
+<ButtonTile
     primaryText={localize('actions.mintNativeToken')}
     secondaryText={localize('general.mintNativeTokenDescription')}
     onClick={onMintNativeTokenClick}
+    icon={IconName.CoinSwap}
 />
