+++ packages/desktop/views/dashboard/developer/components/MintNftButton.svelte
@@ -1,5 +1,6 @@
 <script lang="ts">
-    import { OnboardingButton } from '@ui'
+    import { IconName } from '@bloomwalletio/ui'
+    import { ButtonTile } from '../../../components'
     import { localize } from '@core/i18n'
     import { resetMintNftDetails } from '@core/wallet'
     import { openPopup, PopupId } from '@desktop/auxiliary/popup'
@@ -12,8 +13,9 @@
     }
 </script>
 
-<OnboardingButton
+<ButtonTile
     primaryText={localize('actions.mintNft')}
     secondaryText={localize('general.mintNftDescription')}
     onClick={onMintNftClick}
+    icon={IconName.Image}
 />
