+++ packages/desktop/views/onboarding/components/index.ts
@@ -1,3 +1,2 @@
 export { default as MnemonicInput } from './MnemonicInput.svelte'
-export { default as OnboardingSelectorTile } from './OnboardingSelectorTile'
 export { default as SuccessSvg } from './SuccessSvg.svelte'
