+++ packages/shared/src/components/buttons/index.ts
@@ -1,4 +1,3 @@
 export { default as AddInputButton } from './AddInputButton.svelte'
-export { default as OnboardingButton } from './OnboardingButton.svelte'
 export { default as ReceiveButton } from './ReceiveButton.svelte'
 export { default as TooltipIcon } from './TooltipIcon.svelte'
